pyproject.toml
[tool.poetry]
name = "sonorous"
version = "0.1.0"
description = ""
authors = ["the_byte_bender <cccefg2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyenet = "^1.3.14"
wxpython = "^4.2.1"
pyogg = {git = "https://github.com/Meatbag93/PyOgg"}
cyal = "^0.3.2"
msgpack = "^1.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
